Change to linear model
Changed over to using F&H's Linear Model for computing the updates, as described in section 3.2 of F&H 2006.

Really the only change is introducing the `dt` function, which just loops up and down the array.

Noticeably a bit slower (maybe around 10% or so).  Still appears to have the exact same numerical stability issue though...