Add truncation step
Added a truncation step, as reflected in the code sample for F&H.

I think it is actually just their way of implementing a step that I have put in somewhere else, but I'm not 100% which step that is right at this moment.

It doesn't really appear to have had any effect, apart from perhaps making the whole thing run sliiiightly longer.